Issue #1 - Fixed AR:  Search for short terms (less than 5 ch.) is not initiated by hit on [Enter] key. Using magnifier icon - search is initiated for any length terms.
Issue #2 - Fixed
AR: On Events and Training page - term is not kept within the field and shown only as active filters indication.

Issue #6 - Fixed
AR: After additional verification concluded that "Black Hat Europe How serious are social engineering and human hacking-INT005-000067" is a Research feed item as it is in relation with Ovum brand and could be opened only on Research page.
Also if you go to the Research page and initiate search by same term "engineer" - suggestion list will contain mostly the list of Media feed items and you will not be able to open them